float instead of int

+++ app/controllers/verification_controller.rb
@@ -11,8 +11,10 @@ class VerificationController < ApplicationController
   	responce = RestClient.post vk_api_uri, :ID=>@card_id ,:file1 => passphrase
   	doc = Nokogiri::XML(responce)
 
+
+
   	if doc.xpath('EnrollVerify').attr('Status').text == 'OK'
-      if doc.xpath('/EnrollVerify/CardCompareResult/VoiceKeyScore').text.to_i > 0.7 
+      if doc.xpath('/EnrollVerify/CardCompareResult/VoiceKeyScore').text.to_f > 0.7 
   		  redirect_to :action => 'success', :format => params[:format]	
       else
         redirect_to :action => 'error', :format => params[:format]
